[Feat] Invoice schema refinement

+++ fpt_midterm_pos/src/main/java/com/example/fpt_midterm_pos/dto/InvoiceDTO.java
@@ -1,20 +1,20 @@
 package com.example.fpt_midterm_pos.dto;
 
-import com.example.fpt_midterm_pos.data.model.Customer;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
-
 import java.util.Date;
 import java.util.UUID;
+
+import com.example.fpt_midterm_pos.data.model.Customer;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class InvoiceDTO {
     private UUID id;
-    private int amount;
+    private Double amount;
     private Date date;
     private Date createdAt;
     private Date updatedAt;
